% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{add_surv_time}
\alias{add_surv_time}
\title{Add Segment Start and End Times for Survival Analysis}
\usage{
add_surv_time(grouped_df)
}
\arguments{
\item{grouped_df}{A grouped dataframe containing segment_start and segment_end columns}
}
\value{
A grouped dataframe with additional start_time and end_time columns
}
\description{
Add segment start and end times to be relative to the minimum
start time within each group. Handles edge cases where the end time equals 0
or equals start time by adding 1 to play nice with survival analysis estimators.
}
