% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{adjust_dates}
\alias{adjust_dates}
\title{Adjust Dates in Drug Episode Data}
\usage{
adjust_dates(drug_episodes)
}
\arguments{
\item{drug_episodes}{A data.frame of drug episode-level information with survival events.}
}
\value{
A data.frame of progression dates adjusted to prevent patients from progressing within a week
}
\description{
Adjust Dates in Drug Episode Data
}
