% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{build_off_from_on}
\alias{build_off_from_on}
\title{Build off-treatment from on-treatment}
\usage{
build_off_from_on(on_treatment, lines)
}
\description{
Build off-treatment from on-treatment
}
