% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{construct_state_episodes}
\alias{construct_state_episodes}
\title{Helper Functions and Main Constructor for State Episodes}
\usage{
construct_state_episodes(drug_episodes)
}
\arguments{
\item{drug_episodes}{A data.frame of drug episode-level information with survival events.}

\item{prog_on_cutoff}{Integer. Maximum negative offset allowed between progression and lineenddate.}

\item{death_cutoff, progression_cutoff, next_line_cutoff}{Numeric thresholds used to define transitions.}
}
\value{
A data.frame of structured drug transitions.
}
\description{
These functions help create cleaned and structured multi-state transition datasets
from real-world treatment episode data, suitable for survival analysis or cost-effectiveness modeling.
}
