% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_table.R
\name{death_table}
\alias{death_table}
\title{Generate a Monthly State and Death Transition Table}
\usage{
death_table(drug_transitions, flatiron = TRUE)
}
\arguments{
\item{drug_transitions}{A data frame with patient state transitions.}

\item{flatiron}{Logical, whether to join with Flatiron demographics (default = TRUE).}
}
\value{
A data frame summarizing transitions by month relative to treatment.
}
\description{
Creates a monthly summary table showing number of patients in different states
stratified by gender, and includes death transitions per state.
}
