% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{fix_death_before_progression}
\alias{fix_death_before_progression}
\title{Fix Deaths Before Progression}
\usage{
fix_death_before_progression(lines)
}
\description{
Fix Deaths Before Progression
}
