% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{fix_line_gaps}
\alias{fix_line_gaps}
\title{Make sure there are no gaps between line dates}
\usage{
fix_line_gaps(drug_transitions)
}
\description{
Make sure there are no gaps between line dates
}
