% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%m-\%}
\alias{\%m-\%}
\title{Lubridate operator to subtract time}
\usage{
lhs \%m-\% rhs
}
\arguments{
\item{lhs}{A date-time object.}

\item{rhs}{A period or duration object.}
}
\value{
The result of subtracting \code{rhs} from \code{lhs}.
}
\description{
See \code{lubridate::\link[lubridate]{\%m-\%}} for details.
}
\keyword{internal}
