% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_drug_episodes.R
\name{prep_episode_data}
\alias{prep_episode_data}
\title{Prepare Drug Episode Data from Flatiron drug episodes data}
\usage{
prep_episode_data(
  tumour = "nsclc",
  treatment = "cisplatin|carboplatin",
  drug_separator = ",",
  overlap_threshold = 1,
  exact = FALSE
)
}
\arguments{
\item{tumour}{Character. Tumour type to load. Must match the drug episodes files in the Flatiron exploratory round folder. (e.g., "nsclc").}

\item{treatment}{Character. Regex string of drug names to filter for (e.g., "cisplatin|carboplatin").}

\item{drug_separator}{A character string separating multiple drugs in \code{linename}. Default is ",".}

\item{overlap_threshold}{Integer: number of shared drugs required to collapse lines. Default = 1.}

\item{exact}{boolean: to specify whether we want an exact match to the drug name or a combination therapy}
}
\value{
A cleaned and filtered \code{data.frame} of drug episodes with progression, death, and follow-up dates integrated
}
\description{
This function loads, filters, and processes drug episode data for cancer patients,
focusing on advanced or metastatic treatments. It cleans and
merges treatment episodes, filters based on the specified treatment given at line 1, 2 or 3, and integrates
progression, mortality and last contact data.
}
\examples{
\dontrun{
  episodes <- prep_episode_data(tumour = "nsclc", treatment = "cisplatin|carboplatin")
  head(episodes)
}
}
