% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_episodes_pipeline.R
\name{run_episodes_pipeline}
\alias{run_episodes_pipeline}
\title{Run pipeline with custom tumour definitions}
\usage{
run_episodes_pipeline(tumours, dir = "")
}
\arguments{
\item{tumours}{A tibble with columns "tumour", "treatment", and "exact"}

\item{dir}{Path to directory where the RDS file should be saved.}
}
\description{
Run pipeline with custom tumour definitions
}
