% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{build_on_treatment}
\alias{build_on_treatment}
\title{Build On-Treatment State}
\usage{
build_on_treatment(lines, death_cutoff, progression_cutoff, next_line_cutoff)
}
\description{
Build On-Treatment State
}
