% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{build_progression_from_on}
\alias{build_progression_from_on}
\title{Build progression from on-treatment state}
\usage{
build_progression_from_on(on_treatment, lines)
}
\description{
Build progression from on-treatment state
}
