% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_state_episodes.R
\name{fix_death_before_progression}
\alias{fix_death_before_progression}
\title{Adjust incorrect death times}
\usage{
fix_death_before_progression(lines)
}
\arguments{
\item{lines}{formatted drug episodes data}
}
\value{
adjusted death dates dataframe
}
\description{
Adjust incorrect death times
}
